Patch 14: col_nnz type safety + MTX round-trip & Seurat tests

+++ pyscx/src/lazy_transform.rs
@@ -1036,7 +1036,7 @@ impl ScxLazyTransformedDataset {
                             .backed
                             .col_nnz_masked(kept)
                             .map_err(|e| PyRuntimeError::new_err(e.to_string()))?;
-                        f_counts.iter().map(|&v| v as i64).collect::<Vec<i64>>()
+                        f_counts.iter().map(|&v| v as u32).collect::<Vec<u32>>()
                     }
                     (None, None) => self
                         .backed
@@ -1062,13 +1062,13 @@ impl ScxLazyTransformedDataset {
                     let nnz =
                         crate::projected_agg::col_nnz_masked_projected(&self.backed, kept, cols)
                             .map_err(|e| PyRuntimeError::new_err(e.to_string()))?;
-                    let total: i64 = nnz.iter().sum();
+                    let total: u32 = nnz.iter().sum();
                     Ok((total as usize).into_pyobject(py)?.into_any())
                 }
                 (Some(cols), None) => {
                     let nnz = crate::projected_agg::col_nnz_projected(&self.backed, cols)
                         .map_err(|e| PyRuntimeError::new_err(e.to_string()))?;
-                    let total: i64 = nnz.iter().sum();
+                    let total: u32 = nnz.iter().sum();
                     Ok((total as usize).into_pyobject(py)?.into_any())
                 }
                 (None, Some(kept)) => {
@@ -1115,12 +1115,12 @@ impl ScxLazyTransformedDataset {
             (Some(cols), Some(kept)) => {
                 let nnz = crate::projected_agg::col_nnz_masked_projected(&self.backed, kept, cols)
                     .map_err(|e| PyRuntimeError::new_err(e.to_string()))?;
-                Ok(nnz.iter().sum::<i64>() as usize)
+                Ok(nnz.iter().sum::<u32>() as usize)
             }
             (Some(cols), None) => {
                 let nnz = crate::projected_agg::col_nnz_projected(&self.backed, cols)
                     .map_err(|e| PyRuntimeError::new_err(e.to_string()))?;
-                Ok(nnz.iter().sum::<i64>() as usize)
+                Ok(nnz.iter().sum::<u32>() as usize)
             }
             (None, Some(kept)) => {
                 let all_nnz = self
